# Conflicts:
#	h2/src/main/org/h2/mvstore/DataUtils.java
#	h2/src/main/org/h2/mvstore/MVStore.java
#	h2/src/main/org/h2/mvstore/Page.java

+++ h2/src/main/org/h2/value/ValueByte.java
@@ -130,7 +130,7 @@ public class ValueByte extends Value {
 
     @Override
     public Object getObject() {
-        return Byte.valueOf(value);
+        return value;
     }
 
     @Override
